pyproject.toml
[project]
name = "traffic-ai"
version = "1.0.0"
description = "Traffic signal control RL environment — ScalarX x Meta Hackathon 2026"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"

dependencies = [
    "openenv-core>=0.1.0",
    "stable-baselines3[extra]>=2.3.0",
    "gymnasium>=0.29.0",
    "torch>=2.2.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
